backend/requirements.txt
fastapi
uvicorn
pydantic
psutil
PyJWT
